Fix Failure Loop When Install Gets Run Twice
Made sure to copy the Lucky binary only once, just in case something
causes the install hook to run twice.

+++ src/cli/charm/build/install-hook-template.sh
@@ -36,8 +36,10 @@ if [ ! -f ./bin/lucky ]; then
 # If Lucky was bundled
 else
     # Copy lucky binary to the data dir
-    mkdir -p $bin_dir
-    cp ./bin/lucky $lucky
+    if [ ! -f $lucky ]; then
+        mkdir -p $bin_dir
+        cp ./bin/lucky $lucky
+    fi
 fi
 
 # Start the Lucky daemon
